requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx>=0.25.0
gunicorn==21.2.0
aiofiles==23.2.0
supabase>=2.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
boto3>=1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
